requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
